Implement authentication service, dialogs and better network thingy.
