+++ client/test.html
@@ -1,32 +1,94 @@
 <!doctype html>
 <html lang="en">
-    <head>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 
-        <title>NodeGame</title>
-        <script src="https://code.createjs.com/easeljs-0.8.1.min.js"></script>
-        <script type="text/javascript" src="js/objects.js"></script>
-        <script type="text/javascript" src="js/sockets.js"></script>
-        <script type="text/javascript" src="js/connect.js"></script>
+<title>NodeGame</title>
+<script src="https://code.createjs.com/easeljs-0.8.1.min.js"></script>
+<script type="text/javascript" src="js/authenticate.js"></script>
 
-        <style>
-            html, body {
-                width:  100%;
-                height: 100%;
-                margin: 0px;
-                overflow: hidden;
-                padding: 0px;
-                
-            }
-            
-            #status {
-            	background: #000; border: 1px solid #ffff00; color: #fff; 
-            	position: fixed; top: 10px; left: 10px; z-index: 999; padding: 4px;
-            	font-family: sans-serif; font-size: small;
-            }
-        </style>
-    </head>
-    <body onload="init();">
-        <canvas id="canvas" style="background-color: graytext;"></canvas>
-    </body>
+<script type="text/javascript" src="js/sockets.js"></script>
+<script type="text/javascript" src="js/objects.js"></script>
+<script type="text/javascript" src="js/connect.js"></script>
+<script type="text/javascript" src="js/client.js"></script>
+
+<style>
+html, body {
+	width: 100%;
+	height: 100%;
+	margin: 0px;
+	overflow: hidden;
+	padding: 0px;
+}
+
+#status {
+	background: #000;
+	border: 1px solid #ffff00;
+	color: #fff;
+	position: fixed;
+	top: 10px;
+	left: 10px;
+	z-index: 999;
+	padding: 4px;
+	font-family: sans-serif;
+	font-size: small;
+}
+
+#overlay {
+	position: fixed;
+	top: 0px;
+	left: 0px;
+	z-index: 2;
+	background-color: rgba(0, 0, 0, 0.5);
+	width: 100%;
+	height: 100%;
+}
+
+#authenticator {
+	width: 450px;
+	height: 150px;
+	background: #fff;
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	margin-left: -225px;
+	margin-top: -75px;
+	z-index: 999;
+	padding: 10px;
+}
+
+#authenticator h1 {
+	text-align: center;
+}
+
+#authenticator table, #authenticator input {
+	width: 100%;
+}
+
+#authenticator td {
+	width: 50%;
+}
+</style>
+</head>
+<body onload="init();">
+	<div id="overlay" style="display: none;"></div>
+	<div id="authenticator" style="display: none;">
+		<h1>Logo image here!</h1>
+		<table>
+			<tr>
+				<td>Username</td>
+				<td><input type="text" id="username"></td>
+			</tr>
+			<tr>
+				<td>Password</td>
+				<td><input type="password" id="password"></td>
+			</tr>
+			<tr>
+				<td colspan="2"><input type="button" id="authenticate"
+					value="Log in"></td>
+			</tr>
+		</table>
+	</div>
+	<canvas id="canvas" style="background-color: graytext;"></canvas>
+</body>
 </html>
